f-analytics - Used preserveState rather than if statement when registering store module.  Removed plugin as now just a Service.
